requirements.txt
Flask==3.0.3
mistralai
Pillow
python-dotenv
requests~=2.32.3
pydantic~=2.11.7